Revert "カテゴリをハードコードした"
